+++ packages/core/src/services/microcredit/create.ts
@@ -3,6 +3,8 @@ import { models } from '../../database';
 import { sendNotification } from '../../utils/pushNotification';
 import { NotificationType } from '../../interfaces/app/appNotification';
 import { AppUserModel } from '../../database/models/app/appUser';
+import { MicroCreditFormModel } from '../../database/models/microCredit/form';
+import { BaseError } from '../../utils';
 
 export default class MicroCreditCreate {
     private microCreditContentStorage = new MicroCreditContentStorage();
@@ -73,4 +75,33 @@ export default class MicroCreditCreate {
 
         await sendNotification(users, NotificationType.LOAN_STATUS_CHANGED);
     }
+
+    public saveForm = async (userId: number, form: object, submitted: boolean): Promise<MicroCreditFormModel> => {
+        try {
+            const userForm = await models.microCreditForm.findOrCreate({
+                where: {
+                    userId
+                },
+                defaults: {
+                    form,
+                    userId,
+                    submitted
+                }
+            });
+
+            // update form
+            const newForm = { ...userForm[0].form, ...form };
+
+            // TODO: if submitted, check if all required fields was filled (prismic)
+
+            const data = await userForm[0].update({
+                form: newForm,
+                submitted
+            });
+
+            return data;
+        } catch (error) {
+            throw new BaseError('SAVE_FORM_ERROR', error.message);
+        }
+    };
 }
